@@ -2,9 +2,11 @@
   "name": "angular2-unieconomy-frontend",
   "version": "0.0.1",
   "scripts": {
-    "start": "opn http://localhost:3000 & gulp build.watch.and.serve",
-    "build": "gulp clean & gulp build.dist",
-    "clean": "rimraf dist",
+    "start": "gulp build.dev.watch",
+    "serve": "gulp serve",
+    "buid.dev": "gulp build.dev",
+    "build": "gulp build",
+    "clean": "rimraf dist/**/*",
     "test": "gulp serve.tests",
     "lint": "gulp tslint"
   },
@@ -30,25 +32,35 @@
     "reflect-metadata": "0.1.2",
     "rxjs": "5.0.0-beta.6",
     "systemjs": "0.19.26",
-    "unitable-ng2": "^0.0.0-alpha009",
+    "unitable-ng2": "^0.0.0-alpha016",
     "zone.js": "^0.6.12"
   },
   "devDependencies": {
+    "del": "^2.2.0",
     "gulp": "^3.9.0",
+    "gulp-autoprefixer": "^3.1.0",
     "gulp-concat": "^2.6.0",
     "gulp-connect": "^3.2.0",
+    "gulp-copy": "0.0.2",
+    "gulp-cssmin": "^0.1.7",
     "gulp-inject": "^3.0.0",
     "gulp-live-server": "0.0.29",
     "gulp-load-plugins": "^1.1.0",
     "gulp-plumber": "^1.0.1",
+    "gulp-postcss": "^6.1.1",
     "gulp-rename": "^1.2.2",
+    "gulp-rev": "^7.0.0",
+    "gulp-rev-replace": "^0.4.3",
     "gulp-ruby-sass": "^2.0.5",
     "gulp-run": "^1.6.11",
+    "gulp-sass": "^2.3.1",
+    "gulp-shell": "^0.5.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-template": "^3.0.0",
+    "gulp-tsc": "^1.1.5",
     "gulp-tslint": "^4.3.3",
     "gulp-typescript": "^2.9.2",
-    "gulp-uglify": "^1.5.1",
+    "gulp-uglify": "^1.5.3",
     "gulp-uglifyjs": "^0.6.2",
     "gulp-watch": "^4.3.5",
     "jasmine": "^2.3.2",
@@ -61,16 +73,20 @@
     "karma-mocha": "^0.2.2",
     "karma-mocha-reporter": "^1.2.3",
     "karma-phantomjs-launcher": "^0.2.1",
+    "lite-server": "^2.2.0",
     "live-server": "^0.8.2",
     "mocha": "^2.4.5",
+    "node-sass": "^3.7.0",
     "opn": "^3.0.3",
     "opn-cli": "^1.0.0",
     "phantomjs": "^1.9.18",
     "request": "^2.69.0",
     "rimraf": "^2.4.3",
     "run-sequence": "^1.1.4",
+    "sass-convert": "^0.5.2",
+    "sassdoc": "^2.1.20",
     "tsc": "^1.20150623.0",
     "tslint": "^3.3.0",
-    "typescript": "^1.6.2"
+    "typescript": "^1.8.9"
   }
 }
