@@ -1,104 +1,207 @@
-import {Component, ViewChildren} from '@angular/core';
-import {UniTable, UniTableBuilder, UniTableColumn} from '../../../../../framework/uniTable';
+import {Component, ViewChildren, ViewChild} from '@angular/core';
+import {UniTable, UniTableColumn, UniTableColumnType, UniTableConfig, IContextMenuItem} from 'unitable-ng2/main';
 import {Router} from '@angular/router-deprecated';
 import {UniHttp} from '../../../../../framework/core/http/http';
 import {CustomerInvoiceService} from '../../../../services/services';
-import {CustomerInvoice} from '../../../../unientities';
+import {StatusCodeCustomerInvoice} from '../../../../unientities';
+import {Http, URLSearchParams} from '@angular/http';
+import {AsyncPipe} from '@angular/common';
 
-declare var jQuery;
+import {InvoicePaymentData} from '../../../../models/sales/InvoicePaymentData';
+import {RegisterPaymentModal} from '../../../common/modals/registerPaymentModal';
 
 @Component({
     selector: 'invoice-list',
     templateUrl: 'app/components/sales/invoice/list/invoiceList.html',
-    directives: [UniTable],
-    providers: [CustomerInvoiceService]
+    directives: [UniTable, RegisterPaymentModal],
+    providers: [CustomerInvoiceService],
+    pipes: [AsyncPipe]
 })
-export class InvoiceList {
-    @ViewChildren(UniTable) public tables: any;
 
-    private invoiceTable: UniTableBuilder;
-    private selectedinvoice: CustomerInvoice;
-   
-    constructor(private uniHttpService: UniHttp, private router: Router, private customerInvoiceService: CustomerInvoiceService) {
+export class InvoiceList {
+    @ViewChildren(UniTable) public table: any;
+
+    private invoiceTable: UniTableConfig;
+    private lookupFunction: (urlParams: URLSearchParams) => any;
+
+    @ViewChild(RegisterPaymentModal)
+    private registerPaymentModal: RegisterPaymentModal;
+
+    constructor(private uniHttpService: UniHttp,
+        private router: Router,
+        private customerInvoiceService: CustomerInvoiceService,
+        private http: Http) {
         this.setupInvoiceTable();
     }
 
-    log(err) {
+    private log(err) {
         alert(err._body);
     }
-    
-    createInvoice() {
+
+    public createInvoice() {
         this.customerInvoiceService.newCustomerInvoice().then(invoice => {
             this.customerInvoiceService.Post(invoice)
                 .subscribe(
-                    (data) => {
-                        this.router.navigateByUrl('/sales/invoice/details/' + data.ID);        
-                    },
-                    (err) => { 
-                        console.log('Error creating invoice: ', err);
-                        this.log(err);
-                    }
+                (data) => {
+                    this.router.navigateByUrl('/sales/invoice/details/' + data.ID);
+                },
+                (err) => {
+                    console.log('Error creating invoice: ', err);
+                    this.log(err);
+                }
                 );
-        });           
+        });
+    }
+
+    public onRegisteredPayment(modalData: any) {
+
+        this.customerInvoiceService.ActionWithBody(modalData.id, modalData.invoice, 'payInvoice').subscribe((journalEntry) => {
+            // TODO: Decide what to do here. Popup message or navigate to journalentry ??
+            // this.router.navigateByUrl('/sales/invoice/details/' + invoice.ID);
+            alert('Fakturer er betalt. Bilagsnummer: ' + journalEntry.JournalEntryNumber);
+        }, (err) => {
+            console.log('Error registering payment: ', err);
+            this.log(err);
+        });
     }
 
     private setupInvoiceTable() {
-        var self = this;
+        this.lookupFunction = (urlParams: URLSearchParams) => {
+            let params = urlParams;
 
-        // Define columns to use in the table
-        var invoiceNumberCol = new UniTableColumn('InvoiceNumber', 'Fakturanr', 'string').setWidth('10%');
-       
-        var customerNumberCol = new UniTableColumn('Customer.CustomerNumber', 'Kundenr', 'string')
-            .setNullable(true)
-            .setWidth('10%');
+            if (params == null) {
+                params = new URLSearchParams();
+            }
 
-        var customerNameCol = new UniTableColumn('CustomerName', 'Kunde', 'string');
-
-        var invoiceDateCol = new UniTableColumn('InvoiceDate', 'Fakturadato', 'date')
-            .setFormat('{0: dd.MM.yyyy}')
-            .setWidth('10%');
-
-        var dueDateCol = new UniTableColumn('PaymentDueDate', 'Forfallsdato', 'date')
-            .setFormat('{0: dd.MM.yyyy}')
-            .setWidth('10%');
-
-        var taxInclusiveAmountCol = new UniTableColumn('TaxInclusiveAmount', 'Totalsum', 'number')
-            .setWidth('10%')
-            .setFormat('{0:n}')
-            .setClass('column-align-right');
-
-        var statusCol = new UniTableColumn('StatusCode', 'Status', 'number').setWidth('15%');
-        statusCol.setTemplate((dataItem) => {
-            return this.customerInvoiceService.getStatusText(dataItem.StatusCode); 
-        });
-
-        // Define callback function for row clicks
-        var selectCallback = (selectedItem) => {
-            this.router.navigateByUrl('/sales/invoice/details/' + selectedItem.ID);
+            return this.customerInvoiceService.GetAllByUrlSearchParams(params);
         };
 
-        // Setup table
-        this.invoiceTable = new UniTableBuilder('invoices', false)
-            .setFilterable(false)
-            .setSelectCallback(selectCallback)
-            .setExpand('Customer')
-            .setPageSize(25)
-            .addColumns( invoiceNumberCol, customerNumberCol, customerNameCol, invoiceDateCol, dueDateCol, taxInclusiveAmountCol, statusCol)
-            .setOrderBy('PaymentDueDate','desc')
-            .addCommands({
-                name: 'ContextMenu', text: '...', click: (function (event) {
-                    event.preventDefault();
-                    var dataItem = this.dataItem(jQuery(event.currentTarget).closest('tr'));
+        // Context menu
+        let contextMenuItems: IContextMenuItem[] = [];
+        contextMenuItems.push({
+            label: 'Rediger',
+            action: (rowModel) => {
+                this.router.navigateByUrl(`/sales/invoice/details/${rowModel.ID}`);
+            }
+        });
 
-                    if (dataItem !== null && dataItem.ID !== null) {
-                        self.selectedinvoice = dataItem;
-                        alert('Kontekst meny er under utvikling.');
+        contextMenuItems.push({
+            label: 'Krediter',
+            action: (rowModel) => {
+                this.customerInvoiceService.createCreditNoteFromInvoice(rowModel.ID)
+                    .subscribe((data) => {
+                        this.router.navigateByUrl('/sales/invoice/details/' + data.ID);
+                    },
+                    (err) => {
+                        console.log('Error creating credit note: ', err);
+                        this.log(err);
                     }
-                    else {
-                        console.log('Error in selecting the SupplierInvoices');
-                    }
-                })
-            });
-                               
+                    );
+            },
+            disabled: (rowModel) => {
+                // Possible to credit only if status = Invoiced || PartlyPaid || Paid
+                if (rowModel.StatusCode == StatusCodeCustomerInvoice.Invoiced ||
+                    rowModel.StatusCode == StatusCodeCustomerInvoice.PartlyPaid ||
+                    rowModel.StatusCode == StatusCodeCustomerInvoice.Paid) {
+                    return false;
+                } else {
+                    return true;
+                }
+            }
+        });
+
+        contextMenuItems.push({
+            label: 'Slett',
+            action: (rowModel) => {
+                alert('Delete action - Under construction');
+            },
+            disabled: (rowModel) => {
+                return rowModel['Deleted'];
+            }
+        });
+
+        contextMenuItems.push({
+            label: '-------------',
+            action: () => {}
+        });
+
+        contextMenuItems.push({
+            label: 'Fakturer',
+            action: (rowModel) => {
+                alert('Fakturer action');
+
+                this.customerInvoiceService.Transition(rowModel.ID, rowModel, 'invoice').subscribe(() => {
+                    console.log('== Invoice TRANSITION OK ==');
+                    alert('Fakturert OK');
+
+                    // this.table.refresh(); //TODO Refresh and collect data. Not yet implemented fot uniTable
+                }, (err) => {
+                    console.log('Error fakturerer: ', err);
+                    this.log(err);
+                });
+            },
+            disabled: (rowModel) => {
+                if (rowModel.TaxInclusiveAmount == 0) return true; //Must have saved at minimum 1 item related to the invoice
+                return !rowModel._links.transitions.invoice;
+            }
+        });
+
+        contextMenuItems.push({
+            label: 'Registrer betaling',
+            action: (rowModel) => {
+                const title = `Register betaling, Faktura ${rowModel.InvoiceNumber || ''}, ${rowModel.CustomerName || ''}`;
+                const invoiceData: InvoicePaymentData = {
+                    Amount: rowModel.RestAmount,
+                    PaymentDate: new Date()
+                };
+
+                this.registerPaymentModal.openModal(rowModel.ID, title, invoiceData);
+            },
+            disabled: (rowModel) => {
+                return !rowModel._links.transitions.pay;
+            }
+        });
+
+        contextMenuItems.push({
+            label: 'Skriv ut',
+            action: (rowModel) => {
+                alert('Skriv ut action - Under construction');
+            }
+        });
+
+        // Define columns to use in the table
+        var invoiceNumberCol = new UniTableColumn('InvoiceNumber', 'Fakturanr', UniTableColumnType.Text).setWidth('10%');
+        var customerNumberCol = new UniTableColumn('Customer.CustomerNumber', 'Kundenr', UniTableColumnType.Text).setWidth('10%');
+        var customerNameCol = new UniTableColumn('CustomerName', 'Kundenavn', UniTableColumnType.Text);
+
+        var invoiceDateCol = new UniTableColumn('InvoiceDate', 'Fakturadato', UniTableColumnType.Date).setWidth('10%');
+        var dueDateCol = new UniTableColumn('PaymentDueDate', 'Forfallsdato', UniTableColumnType.Date).setWidth('10%');
+
+        var taxInclusiveAmountCol = new UniTableColumn('TaxInclusiveAmount', 'Totalsum', UniTableColumnType.Number)
+            .setWidth('10%')
+            .setFormat('{0:n}')
+            .setCls('column-align-right');
+
+        var restAmountCol = new UniTableColumn('RestAmount', 'Restsum', UniTableColumnType.Number)
+            .setWidth('10%')
+            .setFormat('{0:n}')
+            .setCls('column-align-right');
+
+        var creditedAmountCol = new UniTableColumn('CreditedAmount', 'Kreditert', UniTableColumnType.Number)
+            .setWidth('10%')
+            .setFormat('{0:n}')
+            .setCls('column-align-right');
+
+        var statusCol = new UniTableColumn('StatusCode', 'Status', UniTableColumnType.Number).setWidth('15%');
+        statusCol.setTemplate((dataItem) => {
+            return this.customerInvoiceService.getStatusText(dataItem.StatusCode, dataItem.InvoiceType);
+        });
+
+        // Setup table
+        this.invoiceTable = new UniTableConfig(false, true)
+            .setPageSize(25)
+            .setColumns([invoiceNumberCol, customerNumberCol, customerNameCol, invoiceDateCol, dueDateCol,
+                taxInclusiveAmountCol, restAmountCol, creditedAmountCol, statusCol])
+            .setContextMenu(contextMenuItems);
     }
 }
